requirements.txt
app==0.0.1
azure-ai-textanalytics==5.1.0b4
azure-common==1.1.26
azure-core==1.10.0
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cors==0.9.7
cycler==0.10.0
cymem==2.0.5
en-core-web-sm==2.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
future==0.18.2
gevent==21.1.0
greenlet==1.0.0
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
msrest==0.6.19
murmurhash==1.0.5
nltk==3.3
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
Pillow==8.0.1
plac==1.1.3
preshed==3.0.5
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
rope==0.18.0
six==1.15.0
spacy==2.3.5
srsly==1.0.5
thinc==7.4.5
tldextract==3.1.0
tqdm==4.56.0
urllib3==1.26.2
wasabi==0.8.0
Werkzeug==1.0.1
wordcloud==1.8.1
zope.event==4.5.0
zope.interface==5.2.0
